Run apt-cacher as PID 1 and allow easy access to other commands
Previously the apt-cacher did not run as PID 1 and therefore did not
receive SIGTERM. This means that the user had to wait for the timeout
and apt-cacher could not do a clean shutdown as it got a SIGKILL.

Not apt-cacher is started via "exec" which allows for a fast and clean
shutdown. Also you can add a command to the "docker run" command, e.g.
"bash", which will be executed instead of apt-cacher. This allows for
easy admin access to the container.